refactor: 예시 컴포넌트 examples 폴더로 이동 및 import 경로 수정, 컨벤션 정비
